Review feedback and more tweaks
