+++ src/System.Private.CoreLib/shared/System/Runtime/InteropServices/ComEventsHelpers.NoCom.cs
@@ -2,9 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-// Dummy implementations of non-portable interop methods that just throw PlatformNotSupportedException
-
-using System.Reflection;
 using System.Runtime.InteropServices.ComTypes;
 
 namespace System.Runtime.InteropServices
